Moving list test sources to within list folder
As tests are growing bigger, its necessary to get a better
folder structure. As for further timer tests that are within
timer folder, list tests sources are being moved to whithin
list folder.
