+++ run_tests.bat
@@ -69,15 +69,15 @@ REM Tests compilation
 %TOOLCHAIN_PATH%\%C_COMPILER% -c tests\main.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\main.o
 %TOOLCHAIN_PATH%\%C_COMPILER% -c tests\mockups\sdl_mockups.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\sdl_mockups.o
 
-%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list_test_setups.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_test_setups.o
-%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list_test_utils.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_test_utils.o
-%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_tests.o
-%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list_create_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_create_tests.o
-%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list_destroy_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_destroy_tests.o
-%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list_getter_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_getter_tests.o
-%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list_find_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_find_tests.o
-%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list_insert_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_insert_tests.o
-%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list_remove_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_remove_tests.o
+%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list\list_test_setups.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_test_setups.o
+%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list\list_test_utils.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_test_utils.o
+%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list\list_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_tests.o
+%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list\list_create_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_create_tests.o
+%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list\list_destroy_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_destroy_tests.o
+%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list\list_getter_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_getter_tests.o
+%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list\list_find_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_find_tests.o
+%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list\list_insert_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_insert_tests.o
+%TOOLCHAIN_PATH%\%C_COMPILER% -c tests\utils\datastructure\list\list_remove_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\list_remove_tests.o
 
 %TOOLCHAIN_PATH%\%C_COMPILER% -c tests\core\system\timer\timer_test_setups.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\timer_test_setups.o
 %TOOLCHAIN_PATH%\%C_COMPILER% -c tests\core\system\timer\timer_tests.c %TEST_INCLUDE_PATHS% %TEST_COMPILATION_PARAMS% -o build\tests\%BUILD_MODE%\timer_tests.o
