+++ tests/utils/datastructure/list/list_insert_tests.c
@@ -2,9 +2,9 @@
 
 #include "utils/datastructure/list.h"
 #include "utils/datastructure/list_private.h"
-#include "utils/datastructure/list_tests.h"
-#include "utils/datastructure/list_test_setups.h"
-#include "utils/datastructure/list_test_utils.h"
+#include "utils/datastructure/list/list_tests.h"
+#include "utils/datastructure/list/list_test_setups.h"
+#include "utils/datastructure/list/list_test_utils.h"
 
 DEFINE_STANDALONE_TEST_FUNC(listPushFrontNullPtr) {
     munit_assert_long(listPushFront(NULL, NULL), ==, LIST_FAIL);
