+++ tests/utils/datastructure/list/list_test_setups.c
@@ -1,4 +1,4 @@
-#include "utils/datastructure/list_test_setups.h"
+#include "utils/datastructure/list/list_test_setups.h"
 
 #include <stdlib.h>
